having some issues with deserializing containers
When the compiler is trying to create templated versions for my deserialization functions, it trys to use the deserialize_container<T>() method on id_types, which gives errors. Even though I have a check to see if the value is even an array or not. I think the fix is constexpr but im not too familar with it